backend/requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioredis==1.3.1
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
astroid==3.2.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==24.2.0
autobahn==23.1.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
bs4==0.0.2
celery==5.4.0
certifi==2024.7.4
cffi==1.17.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==43.0.0
daphne==3.0.2
dataclasses-json==0.6.7
decorator==5.1.1
dill==0.3.8
distro==1.9.0
Django==4.1
django-localflavor==4.0
django-localflavor-us==1.1
django-multiselectfield==0.1.13
django-rest-framework==0.1.0
djangorestframework==3.15.1
exceptiongroup==1.2.2
executing==2.0.1
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
hiredis==3.0.0
httpcore==1.0.5
httpx==0.27.2
hyperlink==21.0.0
idna==3.7
incremental==24.7.2
ipdb==0.13.13
ipython==8.26.0
isort==5.13.2
jedi==0.19.1
jiter==0.5.0
jsonpatch==1.33
jsonpointer==3.0.0
kombu==5.3.7
langchain==0.2.14
langchain-community==0.2.12
langchain-core==0.2.35
langchain-ollama==0.1.1
langchain-text-splitters==0.2.2
langsmith==0.1.105
localflavor==1.9
marshmallow==3.22.0
matplotlib-inline==0.1.7
mccabe==0.7.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.24.4
ollama==0.3.1
openai==1.43.0
orjson==3.10.7
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
platformdirs==4.2.2
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylint==3.2.5
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-stdnum==1.20
pytz==2024.1
PyYAML==6.0.2
redis==5.0.7
requests==2.32.3
service-identity==24.1.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.32
sqlparse==0.5.0
stack-data==0.6.3
stripe==10.6.0
tenacity==8.5.0
tomli==2.0.1
tomlkit==0.13.0
tqdm==4.66.5
traitlets==5.14.3
Twisted==24.3.0
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.9.4
zope.interface==7.0.1

youtube_transcript_api==0.6.2
django-cors-headers==4.4.0
PyMuPDF==1.24.10 
